Refactor synthesis task models to break circular import
